+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 1.1.0
+## [1.1.1] - 2024-04-12
+### Modified
+- Changed how sums are displayed in Savage World trait rolls
+- `/roll` command now displays individual dice results as well as sum
+
+## [1.1.0] - 2023-06-04
 ### Fixed
 - Roll command can define dice by `D` as well as `d`
 
@@ -8,7 +13,7 @@
 - Savage Worlds parties
 - Savage Worlds statuses 
 
-## 1.0.0
+## [1.0.0] - 2023-01-31
 ### Added
 - Slash command for Savage World trait rolls
 - Party and character status management for Savage World